EMS-N-07-06.31 - Implements retry and backoff when message sending fails.

+++ src/main/java/com/rafaelsousa/algashop/ordering/infrastructure/adapters/out/messaging/outbox/OutboxDispatcher.java
@@ -10,6 +10,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.support.TransactionTemplate;
 
+import java.time.OffsetDateTime;
 import java.util.List;
 
 @Slf4j
@@ -31,11 +32,55 @@ public class OutboxDispatcher {
 
         if (!batch.isEmpty()) {
             for (OutboxMessage message : batch) {
-                outboxKafkaSender.send(message);
+                if (!isEligible(message)) continue;
 
-                transactionTemplate.executeWithoutResult(
-                        _ -> outboxMessageRepository.deleteMessage(message.getId()));
+                try {
+                    outboxKafkaSender.send(message);
+                    transactionTemplate.executeWithoutResult(
+                            _ -> outboxMessageRepository.deleteMessage(message.getId()));
+                } catch (Exception ex) {
+                    transactionTemplate.executeWithoutResult(_ -> registerFailed(message, ex));
+                }
             }
         }
+    }
+
+    private void registerFailed(OutboxMessage message, Exception ex) {
+        int attempts = message.getAttempts() + 1;
+        OffsetDateTime now = OffsetDateTime.now();
+
+        OffsetDateTime failedAt = null;
+        if (attempts >= outboxProperties.getMaxAttempts()) {
+            failedAt = now;
+        }
+
+        OffsetDateTime nextAttemptAt = now.plus(outboxProperties.getBackoff());
+        String lastError = extractError(ex);
+
+        outboxMessageRepository.registerFailed(
+                message.getId(), attempts, nextAttemptAt, lastError, failedAt);
+
+        if (failedAt != null) {
+            log.error(
+                    "Permanent failed sending message from outbox {} after {} attempts. Manual intervention required.",
+                    message.getId(),
+                    attempts);
+        }
+    }
+
+    private String extractError(Exception ex) {
+        String description;
+        if (ex.getCause() != null) {
+            description = "Error: %s\n Cause:\n %s".formatted(ex.getMessage(), ex.getCause().getMessage());
+        } else {
+            description = "Error: %s".formatted(ex.getMessage());
+        }
+
+        return description;
+    }
+
+    private boolean isEligible(OutboxMessage message) {
+        return message.getFailedAt() == null
+                && !message.getNextAttemptAt().isAfter(OffsetDateTime.now());
     }
 }
